ci: only cache third-party deps if they exist

+++ ci/before_cache.sh
@@ -18,7 +18,7 @@ ccache -s 2>/dev/null || true
 find "${HOME}/.ccache" -name stats -delete
 
 # Update the third-party dependency cache only if the build was successful.
-if ended_successfully; then
+if ended_successfully && [ -d "${DEPS_BUILD_DIR}" ]; then
   # Do not cache downloads.  They should not be needed with up-to-date deps.
   rm -rf "${DEPS_BUILD_DIR}/build/downloads"
   rm -rf "${CACHE_NVIM_DEPS_DIR}"
